refactor(dashboard): split ~1170-line page.tsx into slot components (LX-B1)
Behavior-identical component split so future dashboard surfaces (endowed
progress LX-B12, leagues strip LX-B9b, review strip LX-B4/B6) land as
additive slots instead of contending for one file.

- hooks/use-dashboard-data.ts: useDashboardData + DashboardData /
  CurrentCourse / ActivityItem types, moved verbatim
- components/dashboard/current-courses-section.tsx: enrolled-course grid,
  unenroll flow, completion-mint overlay
- components/dashboard/activity-section.tsx: paginated activity feed +
  explorer links
- components/dashboard/recommended-courses-section.tsx: recommendations
  grid/carousel
- components/dashboard/name-reveal-dialog.tsx: first-login name reveal
- page.tsx: thin composition root (hero Continue card + identity panel
  placement unchanged)

Zero copy changes, zero data-flow changes; smoke tests added for the two
presentational sections.

Closes #563
